+++ testinfra/backend/__init__.py
@@ -48,11 +48,12 @@ def parse_hostspec(hostspec):
         kw["connection"] = url.scheme
         host = url.netloc
         query = urllib.parse.parse_qs(url.query)
-        for key in ('sudo', 'ssl', 'no_ssl', 'no_verify_ssl'):
+        for key in ('sudo', 'ssl', 'no_ssl', 'no_verify_ssl', 'force_ansible'):
             if query.get(key, ['false'])[0].lower() == 'true':
                 kw[key] = True
         for key in ("sudo_user", 'namespace', 'container', 'read_timeout_sec',
-                    'operation_timeout_sec', 'timeout'):
+                    'operation_timeout_sec', 'timeout', 'controlpersist',
+                    'kubeconfig'):
             if key in query:
                 kw[key] = query.get(key)[0]
         for key in (